# docs/openapi/admin.yaml
openapi: 3.0.3
...
info:
  title: voxeltoad — Admin API
  version: 0.1.0
  description: >
    Management-plane REST API for providers, models, routes, plugins, tenants,
    API keys, quotas, usage, and audit. Consumed by the Control Panel UI and the
    generated TypeScript admin SDK.
servers:
  - url: /
    description: Same-origin (admin plane root)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/usage:
    get:
      tags: [usage]
      summary: List usage records (keyset paginated, time-bounded)
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: tenant
          in: query
          required: false
          schema: { type: string }
          description: super-admin only; filter to a single tenant. Ignored for tenant-admin (implicitly own tenant).
        - name: provider
          in: query
          required: false
          schema: { type: string }
          description: Filter to a single upstream provider.
        - name: model
          in: query
          required: false
          schema: { type: string }
          description: Filter to a single model alias.
        - name: format
          in: query
          required: false
          schema: { type: string, enum: [csv] }
          description: Set to "csv" to export as CSV instead of JSON (2000-row cap, per ADR-0028).
      responses:
        '200':
          description: Usage record page (JSON) or CSV export (when format=csv)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsageList' }
            text/csv:
              schema: { type: string, description: "CSV with headers: id,tenant,group_name,api_key_id,provider,model,prompt_tokens,completion_tokens,cost,created_at" }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/usage/summary:
    get:
      tags: [usage]
      summary: Aggregate usage (grouped, summed), by currency
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - name: group_by
          in: query
          required: false
          schema:
            type: string
            enum: [tenant, model, provider, api_key_id, group_name]
          description: Dimension to group aggregates by.
      responses:
        '200':
          description: Usage aggregates
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsageSummaryList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/usage/timeseries:
    get:
      tags: [usage]
      summary: Time-bucketed usage trends (cost/tokens/requests per hour/day/week)
      parameters:
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - name: bucket
          in: query
          required: false
          schema: { type: string, enum: [hour, day, week], default: day }
          description: Time granularity for the buckets.
        - name: provider
          in: query
          required: false
          schema: { type: string }
        - name: model
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Time-bucketed usage series (sparse — empty buckets omitted).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UsageTimeseriesList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---------------------------------------------------------------------------
  # Audit. Attributed by AFFECTED tenant (ADR-0019): tenant-admin sees ops
  # against its tenant — including super-admin actions on it. Reads not audited.
  # ---------------------------------------------------------------------------
...
  /api/v1/trace/sessions/{session_id}:
    get:
      tags: [trace]
      summary: Session trace payloads — chronological request list (summary, no bodies)
      description: >
        Returns the trace_payloads summary rows for a session_id in chronological
        (ASC) order — one row per request, with the summary dimensions
        (model, status, stop_reason, token/message counts) but WITHOUT the large
        JSONB bodies. Pairs with /request-logs/sessions/{session_id} (metadata
        timeline): drill into a single request via /trace/requests/{request_id}
        for the message + raw bodies. Tenant scope applies identically.
      parameters:
        - name: session_id
          in: path
          required: true
          schema: { type: string }
          description: The X-Voxeltoad-Session header value to trace.
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 200, maximum: 500 }
          description: Caps the number of summary rows returned.
      responses:
        '200':
          description: Session trace payload summaries (timeline order).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TraceSessionPayloads' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/trace/requests/{request_id}:
    get:
      tags: [trace]
      summary: Single-request trace detail — messages + raw bodies (AUDITED read)
      description: >
        Returns the full trace_payloads row for a request_id: the normalized
        message array and the raw upstream request/response bodies. This read is
        AUDITED because it exposes prompt/completion plaintext (ADR-0039 §5).
        Tenant scope applies (a tenant-admin cannot read another tenant's
        payload). 404 when no payload was captured for the request_id (capture
        disabled, or the row expired/retention-dropped).
      parameters:
        - name: request_id
          in: path
          required: true
          schema: { type: string }
          description: The gateway-assigned request correlation id (X-Request-Id).
      responses:
        '200':
          description: Full trace payload (summary + messages + raw bodies).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TracePayloadDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/trace/rows/{id}:
    get:
      tags: [trace]
      summary: Trace payload detail by row id (unique primary key)
      description: >
        Returns the full message + raw layers for a single trace_payloads row,
        addressed by its identity primary key. Use this instead of
        /trace/requests/{request_id} when request_id may be duplicated across
        rows (some clients send the same X-Request-Id for every request in a
        session). The row id is unique so each request resolves to its own
        payload. Reading this is AUDITED (ADR-0039 §5). Tenant scope applies.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
          description: The trace_payloads row id (from TracePayloadSummary.id).
      responses:
        '200':
          description: Full trace payload detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TracePayloadDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------------------
  # Gateway-wide behavior settings (trace capture, future otel/rate_limit).
  # Super-admin only. Hot-reloadable: the data plane applies changes from the
  # config snapshot within one poll interval, no restart needed.
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque operator session token from POST /auth/login.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
      description: Opaque keyset cursor from a prior response's next_cursor. Omit for the first page.
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
      description: Max items to return.
    From:
      name: from
      in: query
      required: false
      schema: { type: string, format: date-time }
      description: Inclusive lower bound on created_at (RFC3339).
    To:
      name: to
      in: query
      required: false
      schema: { type: string, format: date-time }
      description: Exclusive upper bound on created_at (RFC3339).
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing/invalid/expired session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not authorized for this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found (or not in caller's tenant)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type]
          properties:
            message: { type: string }
            type:
              type: string
              description: OpenAI-compatible error type.
              enum:
                - invalid_request_error
                - authentication_error
                - permission_error
                - not_found
                - rate_limit_error
                - api_error
    ListEnvelope:
      type: object
      description: Uniform list envelope. next_cursor="" means no further pages.
      required: [data, next_cursor]
      properties:
        data:
          type: array
          items: {}
        next_cursor:
          type: string
...
    UsageRecord:
      type: object
      properties:
        id: { type: integer, format: int64 }
        tenant: { type: string }
        group_name: { type: string }
        api_key_id: { type: string }
        provider: { type: string }
        model: { type: string }
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        cost: { type: integer, format: int64, description: Micro-units. }
        created_at: { type: string, format: date-time }
    UsageSummaryRow:
      type: object
      description: One aggregate bucket. group_key is the value of the group_by dimension.
      properties:
        group_key: { type: string }
        prompt_tokens: { type: integer, format: int64 }
        completion_tokens: { type: integer, format: int64 }
        cost: { type: integer, format: int64, description: Summed micro-units. }
        request_count: { type: integer, format: int64 }

    # --- Audit ---
...
    UsageList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/UsageRecord' } }
    UsageSummaryList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/UsageSummaryRow' } }
    UsageBucket:
      type: object
      description: One time-bucketed aggregate for the timeseries endpoint.
      properties:
        bucket_start: { type: string, format: date-time }
        prompt_tokens: { type: integer, format: int64 }
        completion_tokens: { type: integer, format: int64 }
        cost: { type: integer, format: int64, description: int64 micro-units (ADR-0013). }
        request_count: { type: integer, format: int64 }
    UsageTimeseriesList:
      allOf:
        - $ref: '#/components/schemas/ListEnvelope'
        - type: object
          properties:
            data: { type: array, items: { $ref: '#/components/schemas/UsageBucket' } }
...
    TracePayloadSummary:
      type: object
      description: >
        One trace event as shown in a request-list view — summary dimensions only,
        no JSONB bodies. 1:1 with a request_logs row via request_id.
      properties:
        id: { type: integer, format: int64 }
        request_id: { type: string }
        session_id: { type: string }
        trace_id: { type: string }
        tenant: { type: string }
        provider: { type: string }
        model_requested: { type: string }
        stream: { type: boolean }
        agent_type: { type: string, description: Detected calling agent (claude-code/codex/codebuddy/workbuddy/opencode); empty when unrecognized. }
        status_code: { type: integer, description: Upstream HTTP status (0 if unknown). }
        stop_reason: { type: string, description: Completion finish/stop reason. }
        n_messages: { type: integer, description: Number of messages in the request. }
        n_tool_use: { type: integer, description: Number of tool calls in the request. }
        created_at: { type: string, format: date-time }
    TracePayloadDetail:
      allOf:
        - $ref: '#/components/schemas/TracePayloadSummary'
        - type: object
          description: >
            Full trace payload for a single request: the message + raw layers. The
            messages/request_raw JSONB bodies are emitted verbatim
            (provider-heterogeneous shapes); response_raw is the verbatim upstream
            response body as TEXT (an SSE transcript for streams). Reading this is
            AUDITED (ADR-0039 §5).
          properties:
            messages:
              type: array
              description: Normalized request message array (adapter.Message[]) — the message layer.
              items:
                type: object
                additionalProperties: true
            request_raw:
              type: object
              description: Original client request body — the raw request layer.
              additionalProperties: true
            response_raw:
              type: string
              description: >
                Upstream response body (reassembled SSE transcript for streams) —
                the raw response layer. Stored as TEXT because streaming responses
                are SSE, not JSON.
            error_raw:
              type: string
              description: Upstream error body surfaced on failure (empty on success).
    TraceSessionPayloads:
      type: object
      description: Session trace payload summaries (request list, timeline order).
      properties:
        session_id: { type: string }
        requests:
          type: array
          items: { $ref: '#/components/schemas/TracePayloadSummary' }
          description: Trace summaries in created_at ASC order (timeline order).

    # --- Gateway-wide behavior settings (hot-reloadable knobs). ---
...